requirements.txt
Django==3.2.5
Pillow==9.0.0
opencv-python==4.5.5.64
opencv-contrib-python==4.5.5.64
numpy==1.22.1
gunicorn==20.1.0
